+++ src/Controller/InvoiceEntryController.php
@@ -15,6 +15,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 #[Route('/admin/invoices/{invoice}/entries')]
@@ -31,6 +32,7 @@ class InvoiceEntryController extends AbstractController
      * @throws EconomicsException
      */
     #[Route('/new/{type}', name: 'app_invoice_entry_new', methods: ['GET', 'POST'])]
+    #[IsGranted('EDIT', 'invoice')]
     public function new(Request $request, Invoice $invoice, InvoiceEntryTypeEnum $type, InvoiceEntryRepository $invoiceEntryRepository): Response
     {
         if ($invoice->isRecorded()) {
@@ -81,6 +83,7 @@ class InvoiceEntryController extends AbstractController
      * @throws EconomicsException
      */
     #[Route('/{id}/edit', name: 'app_invoice_entry_edit', methods: ['GET', 'POST'])]
+    #[IsGranted('EDIT', 'invoiceEntry')]
     public function edit(Request $request, Invoice $invoice, InvoiceEntry $invoiceEntry, InvoiceEntryRepository $invoiceEntryRepository): Response
     {
         $options = [];
@@ -118,6 +121,7 @@ class InvoiceEntryController extends AbstractController
      * @throws EconomicsException
      */
     #[Route('/{id}', name: 'app_invoice_entry_delete', methods: ['POST'])]
+    #[IsGranted('EDIT', 'invoiceEntry')]
     public function delete(Request $request, Invoice $invoice, InvoiceEntry $invoiceEntry, InvoiceEntryRepository $invoiceEntryRepository): Response
     {
         if ($invoice->isRecorded()) {